+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2022-11-01
+
+- Nothing.
+
+### Cave Story
+
+- Nothing.
+
+### Metroid Dread
+ 
+- Added: New Highly Dangerous Logic setting for enabling situations that may be unrecoverable upon saving.
+
+### Metroid Prime
+
+- Nothing.
+
+### Metroid Prime 2: Echoes
+
+- Nothing.
+
 ## [5.1.0] - 2022-10-01
 
 - Added: You can now view past versions of the presets and revert your preset to it.
@@ -14,6 +34,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed: Discord slash command for FAQ has better usability on mobile.
 - Changed: The parent for a preset is now stored in your preferences, instead of in the preset itself.
 - Fixed: The solver can no longer consider collecting a location a requirement to collecting itself. This is a regression from 4.3.0.
+
+### Discord Bot
+
+- Added: `/website` command that gives instructions to where Randovania's website is.
+- Changed: `/randovania-faq` is now just `/faq`.
+- Changed: `/database-inspect` is now just `/database`.
 
 ### Cave Story